+++ examples/load_in_pinocchio.py
@@ -5,7 +5,7 @@
 # Copyright 2022 Stéphane Caron
 
 """
-Load a robot description, specified from the command line, in Pinocchio.
+Load a robot description selected from the command line in Pinocchio.
 
 This example requires Pinocchio, installed by e.g. `conda install pinocchio`.
 """