+++ examples/show_in_pybullet.py
@@ -5,7 +5,7 @@
 # Copyright 2022 Stéphane Caron
 
 """
-Show a robot description, specified from the command line, using PyBullet.
+Show a robot description selected from the command line using PyBullet.
 
 This example is equivalent to `python -m robot_descriptions show_in_pybullet`.
 It requires PyBullet, which can be installed by `pip install pybullet`.